requirements.txt
streamlit
pymupdf
groq
reportlab
python-dotenv
pytesseract
pdf2image
